+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to `NovaPoshta` will be documented in this file
 
+### 2022-11-07
+- [Add] Добавление метода `saveAddTerm` в `AdditionalService`
+
 ### 2022-11-06
 - [Add] Добавление метода `getPDF` в `InternetDocument`
 - [Add] Добавление возможности распечатывать реестр методом `setThisIsScansheet()`